feat(cli): Extract install from open command
